+++ R/zzz.R
@@ -21,6 +21,7 @@
         "funs", "n", "nOrders",  "path_T1", "path_T3",
         "rewardseq.OFF", "rewardseq.ON",
         "totreward", "trstate", "MarketType","predict_CMSUM", "LABEL_CMSUM",
+        "CUMSUM_PNL", "DFR", "PairGain", "Symbol", "aes",
 
 
         # sample function names
@@ -31,7 +32,7 @@
         "if_else", "lag", "mutate_all", "mutate_if", "slice", "str_detect",
         "predict", "read_rds", "summarise", "write.csv", "write_rds", "ymd_hms",
         "filter_all", "full_join", "bind_rows",  "as.h2o", "bind_cols", "testdir",
-        "h2o.performance", "distinct"
+        "h2o.performance", "distinct", "dev.off", "geom_bar", "geom_line", "ggplot", "ggtitle", "pdf"
       )
     )